# api/openapi.yaml
openapi: 3.1.0
info:
  title: StableAgent API
  version: 0.4.0
  description: StableAgent 是一个以稳定性为核心的生产级通用 Agent 平台。本 API 覆盖任务创建查询、AgentEvent timeline 查询、SSE 实时推送、基础 JWT 鉴权和 tenant 隔离。
servers:
  - url: http://localhost:8080
    description: 本地 API Service
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      summary: 健康检查
      security: []
      responses:
        "200":
          description: 服务可用
  /livez:
    get:
      summary: 存活检查
      security: []
      responses:
        "200":
          description: 进程存活
  /readyz:
    get:
      summary: 就绪检查
      security: []
      responses:
        "200":
          description: 服务就绪
  /api/v1/service-info:
    get:
      summary: API 服务信息
      security: []
      responses:
        "200":
          description: 返回当前服务名称和职责
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceInfoResponse"
  /api/v1/tasks:
    post:
      summary: 创建 Agent 任务
      description: 从 JWT 上下文读取 tenant_id 和 user_id，不信任请求体中的租户信息。创建成功后同事务写入 agent_tasks、agent_states、TASK_CREATED event 和 START_WORKFLOW outbox。
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: 任务已创建，状态为 QUEUED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/Unavailable"
    get:
      summary: 查询任务列表
      description: 默认按 created_at desc、task_id desc 稳定排序。admin/owner 可查看租户内任务，普通 member 只查看自己创建的任务。
      operationId: listTasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - created_at_desc
            default: created_at_desc
      responses:
        "200":
          description: 任务分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTasksResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/tasks/{task_id}:
    get:
      summary: 查询任务详情
      description: 查询始终携带 tenant_id；任务存在但当前用户不是创建者且不是 admin/owner 时返回 403。
      operationId: getTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
      responses:
        "200":
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/tasks/{task_id}/events:
    get:
      summary: 查询任务 timeline
      description: 按 event_id 升序读取任务事件。支持 after_event_id 游标、limit 和 type 过滤，查询始终携带 tenant_id 并复用任务读取权限。
      operationId: listTaskEvents
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - name: after_event_id
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AgentEventType"
      responses:
        "200":
          description: 任务事件列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTaskEventsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/tasks/{task_id}/events/stream:
    get:
      summary: 订阅任务事件 SSE
      description: 连接建立后先按 Last-Event-ID 补发历史事件，再订阅实时事件。服务端写入 id、event 和 data 字段，事件名固定为 agent_event。
      operationId: streamTaskEvents
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: integer
            minimum: 0
        - name: last_event_id
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: SSE 事件流
          content:
            text/event-stream:
              schema:
                type: string
                example: "id: 12\nevent: agent_event\ndata: {\"event_id\":12,\"type\":\"STEP_COMPLETED\",\"task_id\":\"task_01JZK9X2R6Z8M9\"}\n\n"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT。payload 必须包含 tenant_id，以及 user_id 或 sub；role 可选，admin/owner 可查看租户任务，默认 member。
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未认证或 JWT 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 已认证但无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unavailable:
      description: 依赖服务不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ServiceInfoResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - service
            - role
          properties:
            service:
              type: string
            role:
              type: string
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required:
        - goal
        - budget
      properties:
        goal:
          type: string
          minLength: 1
          maxLength: 8000
          example: 分析需求文档并生成生产级 Go 技术方案
        budget:
          $ref: "#/components/schemas/Budget"
        constraints:
          type: object
          additionalProperties: true
          example:
            language: zh-CN
            require_approval_for_high_risk_tools: true
    Budget:
      type: object
      additionalProperties: false
      required:
        - max_steps
        - max_tokens
        - max_tool_calls
        - max_cost_usd
      properties:
        max_steps:
          type: integer
          minimum: 1
          maximum: 1000
        max_tokens:
          type: integer
          minimum: 1
          maximum: 10000000
        max_tool_calls:
          type: integer
          minimum: 1
          maximum: 10000
        max_cost_usd:
          type: number
          exclusiveMinimum: 0
          maximum: 100000
    BudgetUsage:
      type: object
      required:
        - used_steps
        - used_tokens
        - used_tool_calls
        - used_cost_usd
      properties:
        used_steps:
          type: integer
        used_tokens:
          type: integer
        used_tool_calls:
          type: integer
        used_cost_usd:
          type: number
    TaskStatus:
      type: string
      enum:
        - QUEUED
        - RUNNING
        - WAITING_APPROVAL
        - PAUSED
        - CANCELING
        - CANCELED
        - SUCCEEDED
        - FAILED
    AgentEventType:
      type: string
      enum:
        - TASK_CREATED
        - TASK_STARTED
        - TASK_COMPLETED
        - TASK_FAILED
        - TASK_CANCELLED
        - TASK_RESUMED
        - STEP_STARTED
        - STEP_COMPLETED
        - STEP_FAILED
        - LLM_CALL_STARTED
        - LLM_CALL_COMPLETED
        - TOOL_CALL_STARTED
        - TOOL_CALL_COMPLETED
        - TOOL_APPROVAL_REQUIRED
    AgentEvent:
      type: object
      required:
        - event_id
        - task_id
        - type
        - input
        - output
        - metadata
        - created_at
      properties:
        event_id:
          type: integer
        task_id:
          type: string
        type:
          $ref: "#/components/schemas/AgentEventType"
        input:
          type:
            - object
            - array
            - string
            - number
            - boolean
            - "null"
        output:
          type:
            - object
            - array
            - string
            - number
            - boolean
            - "null"
        metadata:
          type: object
          additionalProperties: true
          properties:
            user_id:
              type: string
            request_id:
              type: string
            source:
              type: string
            step_id:
              type: string
            workflow_id:
              type: string
            attempt:
              type: integer
        trace_id:
          type: string
        span_id:
          type: string
        created_at:
          type: string
          format: date-time
    CreatedTask:
      type: object
      required:
        - task_id
        - status
        - created_at
      properties:
        task_id:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        created_at:
          type: string
          format: date-time
    TaskDetail:
      type: object
      required:
        - task_id
        - goal
        - status
        - current_step
        - budget
        - budget_usage
        - created_at
        - updated_at
      properties:
        task_id:
          type: string
        goal:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        current_step:
          type: integer
        budget:
          $ref: "#/components/schemas/Budget"
        budget_usage:
          $ref: "#/components/schemas/BudgetUsage"
        trace_id:
          type: string
        last_error_code:
          type: string
        last_error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskSummary:
      type: object
      required:
        - task_id
        - goal
        - status
        - budget
        - budget_usage
        - created_at
        - updated_at
      properties:
        task_id:
          type: string
        goal:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        budget:
          $ref: "#/components/schemas/Budget"
        budget_usage:
          $ref: "#/components/schemas/BudgetUsage"
        trace_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTaskResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/CreatedTask"
    GetTaskResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/TaskDetail"
    ListTasksResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - items
            - page
            - page_size
            - has_more
            - sort
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/TaskSummary"
            page:
              type: integer
            page_size:
              type: integer
            has_more:
              type: boolean
            next_page:
              type: integer
            sort:
              type: string
              enum:
                - created_at_desc
            status:
              $ref: "#/components/schemas/TaskStatus"
    ListTaskEventsResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - items
            - next_after_event_id
            - has_more
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AgentEvent"
            next_after_event_id:
              type: integer
            has_more:
              type: boolean
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: INVALID_ARGUMENT
            message:
              type: string
              example: goal 不能为空
            request_id:
              type: string